Restructuring graph
